+++ app/backend/app/Http/Requests/Admins/AdminCreateRequest.php
@@ -9,7 +9,7 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Log;
 use App\Http\Requests\BaseRequest;
-use App\Models\Roles;
+use App\Models\Masters\Roles;
 use App\Repositories\Admins\Roles\RolesRepositoryInterface;
 
 class AdminCreateRequest extends BaseRequest
