+++ app/backend/app/Models/Masters/Permissions.php
@@ -5,7 +5,7 @@ namespace App\Models\Masters;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use App\Models\Roles;
+use App\Models\Masters\Roles;
 
 class Permissions extends Model
 {
