+++ app/backend/app/Models/Masters/Roles.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Masters;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use App\Models\Masters\Permissions;
-use App\Models\Roles;
+use App\Models\Masters\Admins;
 
-class RolePermissions extends Model
+class Roles extends Model
 {
     use HasFactory;
     use SoftDeletes;
@@ -16,15 +15,14 @@ class RolePermissions extends Model
     // カラム一覧
     public const ID = 'id';
     public const NAME = 'name';
-    public const SHORT_NAME = 'short_name';
-    public const ROLE_ID = 'role_id';
-    public const PERMISSION_ID = 'permission_id';
+    public const CODE = 'code';
+    public const DETAIL = 'detail';
     public const CRREATED_AT = 'created_at';
     public const UPDATED_AT = 'updated_at';
     public const DELETED_AT = 'deleted_at';
 
     //テーブル名指定
-    protected $table = 'role_permissions';
+    protected $table = 'roles';
 
     // カラムの自動更新をEloquentに許可
     public $timestamps = true;
@@ -41,6 +39,8 @@ class RolePermissions extends Model
 
     // 更新可能なカラムリスト
     protected $fillable = [
+        self::NAME,
+        self::CODE,
         self::UPDATED_AT
     ];
 
@@ -50,27 +50,4 @@ class RolePermissions extends Model
      * @var array
      */
     protected $hidden = [];
-
-    /**
-    * Define a many-to-many relationship.
-    * 設定されているロールの取得
-    *
-    * @return Roles|null
-    */
-    public function roles()
-    {
-        return $this->belongsTo(Roles::class, 'role_id');
-        // return $this->belongsToMany(Roles::class, 'role_id');
-    }
-
-    /**
-    * Define a many-to-many relationship.
-    * 設定されている権限の取得
-    *
-    * @return Permissions|null
-    */
-    public function permissions()
-    {
-        return $this->belongsTo(Permissions::class, 'permission_id');
-    }
 }
