+++ app/backend/app/Repositories/Admins/Permissions/PermissionsRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Repositories\Admins\Permissions;
 
-use App\Models\Roles;
+use App\Models\Masters\Roles;
 use App\Models\Masters\Permissions;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Collection;
