+++ app/backend/app/Repositories/Admins/RolePermissions/RolePermissionsRepository.php
@@ -2,9 +2,9 @@
 
 namespace App\Repositories\Admins\RolePermissions;
 
-use App\Models\Roles;
+use App\Models\Masters\Roles;
 use App\Models\Masters\Permissions;
-use App\Models\RolePermissions;
+use App\Models\Masters\RolePermissions;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Collection;
