+++ app/backend/database/seeders/DatabaseSeeder.php
@@ -2,7 +2,6 @@
 
 namespace Database\Seeders;
 
-use App\Models\RolePermissions;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
@@ -21,8 +20,8 @@ class DatabaseSeeder extends Seeder
         $this->call(Masters\ManufacturersTableSeeder::class);
         $this->call(Masters\ProductsTableSeeder::class);
         $this->call(Masters\PermissionsTableSeeder::class);
-        $this->call(RolesTableSeeder::class);
-        $this->call(RolePermissionsTableSeeder::class);
+        $this->call(Masters\RolesTableSeeder::class);
+        $this->call(Masters\RolePermissionsTableSeeder::class);
         $this->call(Masters\AdminsRolesTableSeeder::class);
         $this->call(UsersTableSeeder::class);
     }
